Initial commit with all project files

+++ AuthService.Application/DTOs/Auth/AuthResultDto.cs
@@ -9,7 +9,7 @@ namespace AuthService.Application.DTOs.Auth
         public int UserId { get; set; }
         public string AccessToken { get; set; }
         public string RefreshToken { get; set; }
-        public int ExpiresIn { get; set; }
+        public double ExpiresIn { get; set; }
         public string TokenType { get; set; }
         public bool IsNewUser { get; set; }
         public string Error { get; set; }
