+++ AuthService.Application/DTOs/Auth/LoginResponseDto.cs
@@ -8,7 +8,7 @@ namespace AuthService.Application.DTOs.Auth
         public bool Success { get; set; }
         public string AccessToken { get; set; } = string.Empty;
         public string RefreshToken { get; set; } = string.Empty;
-        public int ExpiresIn { get; set; }
+        public Double ExpiresIn { get; set; }
         public DateTime ExpiresAt { get; set; }
         public UserInfoDto User { get; set; } = new();
         public RoleTypeEnum Role { get; set; }
